Numpy CAPI example
This example build a python module that gets numpy arrays built in C
using the numpy C-API. I made it to figure out the mystery of NO_IMPORT_ARRAY
and PY_ARRAY_UNIQUE_SYMBOL - very lacking in documentation and hardly an
examples. So here is one.

Build with numpy-1.26.2 and python 3.12.

$ > make

$ > python test.py